+++ src/components/Flags/Flags.js
@@ -1,10 +1,17 @@
 import React from "react";
+import { json } from "react-router-dom";
 
 const Flags = () => {
   const [country, setCountry] = React.useState({});
   const [hotStreak, setHotStreak] = React.useState(
     parseInt(localStorage.getItem("hotStreakFlags"))
       ? parseInt(localStorage.getItem("hotStreakFlags"))
+      : 0
+  );
+
+  const [maxHotStreak, setMaxHotStreak] = React.useState(
+    parseInt(localStorage.getItem("maxHotStreakFlags"))
+      ? parseInt(localStorage.getItem("maxHotStreakFlags"))
       : 0
   );
 
@@ -15,6 +22,11 @@ const Flags = () => {
     if (e.target.id.split(" ")[0] === e.target.id.split(" ")[1]) {
       setHotStreak((e) => e + 1);
       localStorage.setItem("hotStreakFlags", parseInt(hotStreak + 1));
+      if (hotStreak + 1 > maxHotStreak) {
+        setMaxHotStreak(hotStreak + 1);
+        localStorage.setItem("maxHotStreakFlags", parseInt(hotStreak + 1));
+      }
+
       randomCountry();
     } else {
       setHotStreak(0);
@@ -28,6 +40,7 @@ const Flags = () => {
       .then((res) => res.json())
       .then((data) => {
         setCountry(data);
+        localStorage.setItem("flagGame", JSON.stringify(data));
         data.options.push(data.country);
         setError("");
       })
@@ -35,7 +48,11 @@ const Flags = () => {
   };
 
   React.useEffect(() => {
-    randomCountry();
+    if (localStorage.getItem("flagGame")) {
+      setCountry(JSON.parse(localStorage.getItem("flagGame")));
+    } else {
+      randomCountry();
+    }
   }, []);
 
   return (
@@ -45,7 +62,7 @@ const Flags = () => {
           <>
             <div id="hotStreak">
               <img src="/img/fire.svg" alt="hotStreak" />
-              {hotStreak}
+              {hotStreak} {maxHotStreak ? `(${maxHotStreak})` : null}
             </div>
             <div id="img">
               {country ? (